requirements/common.txt
Django==3.2.7
djangorestframework==3.12.4
django-environ==0.3.1
django-cors-headers==3.10.0
certifi
